ui-charts: move TrackOccupancyDiagram types and consts to /lib
