+++ front/ui/ui-charts/src/trackOccupancyDiagram/index.ts
@@ -1,7 +1,7 @@
 import './styles/main.css';
 
-export * from './components/types';
-export * from './components/consts';
+export * from './lib/types';
+export * from './lib/consts';
 export { isOccupancyPickingElement } from './components/layers/OccupancyZonesLayer';
 export { default as TrackOccupancyCanvas } from './components/TrackOccupancyCanvas';
 export { default as TrackOccupancyManchette } from './components/TrackOccupancyManchette';
