+++ front/ui/ui-charts/src/trackOccupancyDiagram/lib/consts.ts
