+++ front/ui/ui-charts/src/trackOccupancyDiagram/lib/types.ts
